Changed tokenStreamIterator see desc.
Make the position in the tokenStreamIterator start at 1 rather than 0
with 0 reserved to indicate the empty position. While this may conflict
with the usual arrays start at 0 way of doing things this allows use to
use unsigned types which give significantly larger position ranges
within the same memory. The position starting at 0 also has little
actual effect because in the planned use case the position is never
directly entered by the programmer and is instead used to go back to a
position previously retrieved using getPos().