Small modifications to tokenStreamIterator + tests
Changed tests to reflect change that position 0 is the empty position.
Also made setPos return false if you try and set the position to the
empty position. Considering removing the existance of an empty position and instead just having the position undefined (return exception when attempt to get) if the tokenStreamIterator is empty.